requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.3
aiohttp==3.9.1
aiosignal==1.3.1
aiosqlite==0.19.0
alembic==1.13.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.10.4
asgiref==3.11.1
astunparse==1.6.3
async-timeout==5.0.1
asyncpg==0.29.0
attrs==24.2.0
Automat==24.8.1
base58==2.1.1
bcrypt==5.0.0
beautifulsoup4==4.12.3
blinker==1.9.0
bson==0.5.10
cachetools==5.5.0
certifi==2026.2.25
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
click-default-group==1.2.4
cloup==2.1.2
cmake==4.0.2
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.3
coverage==7.10.7
cryptography==44.0.1
cssselect==1.2.0
cycler==0.12.1
Cython==3.0.11
dacite==1.8.1
decorator==5.2.1
deepface==0.0.93
defusedxml==0.7.1
Deprecated==1.2.15
Django==5.2.12
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dlib @ https://github.com/z-mahmud22/Dlib_Windows_Python3.x/raw/main/dlib-19.24.1-cp311-cp311-win_amd64.whl#sha256=6f1a5ee167975d7952b28e0ce4495f1d9a77644761cf5720fb66d7c6188ae496
dnspython==2.7.0
docopt==0.6.2
drf-spectacular==0.29.0
ecdsa==0.19.1
email-validator==2.1.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.104.1
ffmpeg-python==0.2.0
filelock==3.17.0
fire==0.7.0
Flask==3.0.0
flask-cors==5.0.1
flatbuffers==24.3.25
fonttools==4.61.0
frozenlist==1.4.1
future==1.0.0
gallery_dl==1.30.5
gast==0.4.0
gdown==5.2.0
glcontext==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.36.0
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.0.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.72.0
greenlet==3.2.3
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==21.2.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httplib2==0.31.0
httptools==0.6.4
httpx==0.25.2
hyperlink==21.0.0
idna==3.10
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
internetarchive==5.4.0
isosurfaces==0.1.0
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
jaconv==0.4.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.0
Js2Py==0.74
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==2.12.0
kiwisolver==1.4.9
libclang==18.1.1
llvmlite==0.43.0
loguru==0.7.2
lxml==5.3.1
lz4==4.4.4
magic-filter==1.0.12
Mako==1.3.10
manim==0.18.0
ManimPango==0.6.1
mapbox_earcut==1.0.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.7.2
mdurl==0.1.2
ml_dtypes==0.5.0
moderngl==5.12.0
moderngl-window==2.4.4
moviepy==1.0.3
mtcnn==1.0.0
multidict==6.1.0
multipledispatch==1.0.0
mutagen==1.47.0
networkx==3.2.1
numba==0.60.0
numpy==1.24.3
oauthlib==3.2.2
opencv-contrib-python==4.11.0.86
opencv-python==4.8.0.76
openpyxl==3.1.5
opt_einsum==3.4.0
orjson==3.11.2
packaging==24.2
pandas==2.0.3
parsel==1.10.0
passlib==1.7.4
pefile==2023.2.7
Pillow==10.0.0
pipwin==0.5.2
platformdirs==4.3.6
pluggy==1.6.0
proglog==0.1.12
propcache==0.2.0
Protego==0.4.0
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg==3.2.4
psycopg2==2.9.10
psycopg2-binary==2.9.9
pyaes==1.6.1
pyarmor==9.2.3
pyarmor.cli.core==8.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.29.0
pycparser==2.22
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
PyDispatcher==2.0.7
pydub==0.25.1
pyglet==2.1.11
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pyjsparser==2.7.1
PyJWT==2.11.0
pykakasi==2.3.0
pyOpenSSL==25.0.0
pyparsing==3.0.9
pypng==0.20220715.0
PyPrind==2.11.3
pyrr==0.10.3
pySmartDL==1.3.4
PySocks==1.7.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.15.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-slugify==8.0.4
python-telegram-bot==20.6
pytube==15.0.0
pytz==2024.2
pywin32-ctypes==0.2.3
PyYAML==6.0.1
qrcode==7.4.2
queuelib==1.7.0
RapidFuzz==3.10.1
redis==5.0.1
referencing==0.37.0
requests==2.32.5
requests-file==2.1.0
requests-oauthlib==2.0.0
retina-face==0.0.17
rich==13.5.2
rpds-py==0.30.0
rsa==4.9
scipy==1.11.1
Scrapy==2.12.0
scrapy-splash==0.11.1
screeninfo==0.8.1
service-identity==24.2.0
six==1.17.0
skia-pathops==0.7.4
sniffio==1.3.1
soundcloud-v2==1.6.0
soupsieve==2.6
spotdl==4.2.11
spotipy==2.23.0
SQLAlchemy==2.0.23
sqlparse==0.5.5
srt==3.5.3
starlette==0.27.0
structlog==25.4.0
svgelements==1.9.6
syncedlyrics==1.0.1
Telethon==1.37.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
text-unidecode==1.3
tldextract==5.1.3
tqdm==4.66.1
Twisted==24.11.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
Unidecode==1.3.8
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.24.0
vosk==0.3.45
w3lib==2.3.1
watchdog==3.0.0
watchfiles==1.0.4
websockets==14.2
Werkzeug==3.1.3
win32_setctime==1.2.0
wrapt==1.14.1
yarl==1.14.0
youtube-search-python==1.6.6
youtube-transcript-api==0.6.3
yt-dlp==2026.1.31
ytmusicapi==1.8.2
zope.interface==7.2
